refactor: voltando o useEffect para o app
