+++ src/context/ContatosProvider.js
@@ -1,7 +1,6 @@
-import React, { useState, useMemo, useEffect } from 'react';
+import React, { useState, useMemo } from 'react';
 import { node } from 'prop-types';
 import ContatosContext from './ContatosContext';
-import { contatosData } from '../mocks/contatosData';
 
 function ContatosProvider({ children }) {
   // API
@@ -39,27 +38,6 @@ function ContatosProvider({ children }) {
   //Gráficos
   const [languageTotals, setLanguageTotals] = useState({});
 
-  useEffect(() => {
-    const fetchData = async () => {
-      const dataFromLocalStorage = localStorage.getItem('contatos');
-      if (dataFromLocalStorage) {
-        const parsedData = JSON.parse(dataFromLocalStorage);
-        return setListaContatos(parsedData);
-      } else {
-        const response = await fetch('https://my.api.mockaroo.com/lestetelecom/test.json?key=f55c4060');
-        if (!response.ok) {
-          setLoading(false);
-          return setListaContatos(contatosData);
-        }
-        const data = await response.json();
-        return setListaContatos(data);
-      }  
-    };
-    
-    fetchData();
-  }, [listaContatos]);
-    
-  
   const values = useMemo(() => ({
     newContato,
     setNewContato,
